Move function schemas into own file
